Added profile pic's and username in commentary section.

+++ app/src/main/java/com/dsoft/liverpooldirectory/data/api/dto/vk/comments/ItemComments.kt
@@ -2,8 +2,8 @@ package com.dsoft.liverpooldirectory.data.api.dto.vk.comments
 
 data class ItemComments(
     val attachments: List<Attachment>?,
-    val date: Int?,
-    val from_id: Int?,
+    val date: Long?,
+    val from_id: Long?,
     val id: Int?,
     val owner_id: Int?,
     val parents_stack: List<Any>?,
